fix: split terms page into server wrapper and client component to fix revalidate prerender error
